feat(商城): 首页
添加金刚位跳转

+++ waynboot-common/src/main/java/com/wayn/common/core/domain/shop/Diamond.java
@@ -45,7 +45,7 @@ public class Diamond extends ShopBaseEntity implements Serializable {
     /**
      * 跳转栏目ID或者分类ID
      */
-    private Integer valueId;
+    private Long valueId;
 
     /**
      * 跳转连接
